Eliminated unnecessary remove() calls in jpeg_data_save_file.
The open truncates the file in the first case anyway, and removing it
first could give the new file to different ownership or permissions. The
second case might actually successfully remove a file that couldn't be
written to due to permissions, which would be an unexpected result.
Found by Coverity.