Switch the Cirrus FreeBSD 11.x build to 11.3 and add a 13.0 build.
Also, select the images using image_family to get the latest snapshots
automatically and increase the git clone depth. If more commits are
submitted to master between the time of triggering the first build and
the time the final build gets started, the desired commit is no longer
at HEAD and the build will error out.
[skip ci]