configure.ac: Add URL parameter to AC_INIT invocation
Add the URL of the libexif homepage to the AC_INIT invocation.

The resulting AC_PACKAGE_URL m4 macro and PACKAGE_URL sh var
is not used anywhere yet, but it might be useful for
documentation purposes.

The AC_INIT URL parameter is supported since at least
Autoconf version 2.69 which we AC_PREREQ.