CPP macros beginning/ending with underscores are reserved
C preprocessor macros beginning/ending with underscores are
reserved for system use, so application software macros are
supposed to be named differently.

For a header file foo-bar.h, this consistently uses a macro
name EXIF__FOO_BAR_H for the usual

        #ifndef EXIF__FOO_BAR_H
        #define EXIF__FOO_BAR_H
        ... content of the header file ...
        #endif /* EXIF__FOO_BAR_H */

Found using "clang -Weverything".

This should work, unless when run with an ancient C preprocessor
which might only the first 8 or 10 characters to distinguish
macro names and ignores the remainder of the name.