Enable CI builds on GitHub Actions
These are replacements for the Travis builds removed in commit 59569a50.
Not all are recreated here, but this brings back the most important
checks that were lost.  This include the code coverage build which
hasn't been updated since the Travis builds stopped over 3 years ago.
The results are pushed to Coveralls at
https://coveralls.io/github/libexif/exif